New programs implemented
* README.md: Corresponding introductions added
* display.py: Optimized in argument parsing
* gen_encoding.py: Optimized
* display_faces.py: Created for putting rectangles on faces
* paint_stuff.py: Allow users to paint green rectangles on pictures

Here are why and how to optimize argument parsing in Python:
As we know, we load a number of modules to in order to simplify the
procedure, but loading process is time-consuming. But Python allows
us to import modules whenever we want. Therefore I let the program
parse the arguments first before loading other large modules, and this
is a time-saving solution for invalid command line argument.

Just a note, though TL;DR